# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CrmApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5194/"
    }
  ],
  "paths": {
...
    "/api/deals/{dealId}/move": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDealRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDealRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/deals/{dealId}/score": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDealResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDealResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDealResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/deals/{dealId}/draft-email": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEmailRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEmailRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDraftResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDraftResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDraftResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/deals/{dealId}": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/deals/{dealId}/summarize": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeDealResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeDealResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeDealResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/deals/{dealId}/next-best-action": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NextBestActionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextBestActionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextBestActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/deals/{dealId}/activities": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogActivityRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogActivityRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LogActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/deals/{id}": {
      "get": {
        "tags": [
          "PublicDeals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityDto": {
        "required": [
          "id",
          "type",
          "body",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "body": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DealDetailDto": {
        "required": [
          "id",
          "title",
          "stageName",
          "amountCents",
          "currency",
          "forecastCategory",
          "aiScore",
          "aiScoreRationale",
          "aiSummary",
          "aiSummarizedAt",
          "activitiesSinceSummary",
          "contactNames",
          "activities",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "stageName": {
            "type": "string"
          },
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "aiScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "aiScoreRationale": {
            "type": [
              "null",
              "string"
            ]
          },
          "aiSummary": {
            "type": [
              "null",
              "string"
            ]
          },
          "aiSummarizedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "activitiesSinceSummary": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "contactNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityDto"
            }
          },
          "customFields": {
            "type": "object"
          }
        }
      },
      "DraftEmailRequest": {
        "required": [
          "instruction"
        ],
        "type": "object",
        "properties": {
          "instruction": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "EmailDraftResponse": {
        "required": [
          "jobId"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          }
        }
      },
...
      "LogActivityRequest": {
        "required": [
          "type",
          "body"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
...
      "MoveDealRequest": {
        "required": [
          "stageId"
        ],
        "type": "object",
        "properties": {
          "stageId": {
            "type": "string"
          }
        }
      },
      "NextBestActionResponse": {
        "required": [
          "jobId"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          }
        }
      },
...
      "PublicDealDto": {
        "required": [
          "id",
          "companyId",
          "pipelineId",
          "stageId",
          "amountCents",
          "currency",
          "forecastCategory",
          "contactIds",
          "closedAt",
          "createdAt",
          "updatedAt",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": [
              "null",
              "string"
            ]
          },
          "pipelineId": {
            "type": "string"
          },
          "stageId": {
            "type": "string"
          },
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "contactIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "closedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "customFields": {
            "type": "object"
          }
        }
      },
...
      "ScoreDealResponse": {
        "required": [
          "jobId"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          }
        }
      },
...
      "SummarizeDealResponse": {
        "required": [
          "jobId"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          }
        }
      },
...
      "UpdateDealRequest": {
        "required": [
          "amountCents",
          "currency",
          "forecastCategory",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "customFields": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
    }
  },
...
}